Add Redis support to Qserv Kafka bridge
Set up an internal Redis to store query state, allowing the bridge
to be restarted without losing track of what queries are running.
Ensure only one copy of the bridge is running at a time, since we
don't currently have a locking mechanism to prevent multiple copies
of the bridge from processing the same Qserv query.