Remove unnecessary debug settings from Prompt Processing.
Some settings were added to the Prompt Processing config in case we
needed to quickly adjust them for first photon. Now that we have a
mostly stable service, they shouldn't be needed. This commit does *not*
remove debug settings related to full-scale runs, as those have not yet
been validated.