DM-51209: atlantis - disable in roundtable-dev
The GitHub atlantis app must be installed in at least on repo for
atlantis to start. Since the roundtable-prod Atlantis is now serving the
one repo we use Atlantis for (prodromos), there are no repos we can
attach the dev Atlantis app to, so let's disable the dev Atlantis
instance.