Fix link check errors and tweak GitHub CI job
Trigger the GitHub linkcheck CI job only on documentation changes,
since it's expensive and triggering it on every `values.yaml` and
`Chart.yaml` change feels like a bit too much. Fix a couple of
broken links, and add a timeout on the link check that's about twice
the current runtime.