noteburst: convert worker to stateful set
Make the worker a StatefulSet instead of a deployment so that every
replica can have a consistent identifier (the StatefulSet index) to use
to reserve an RSP identity. If a worker shuts down uncleanly, a Nublado
lab could remain running when the worker starts up again. Having this
consistent ID lets us confidently shut the lab down upon restart without
the fear that the lab is being used by some other replica.