Update to Qserv Kafka bridge 1.1.0
Add support for quota enforcement for simultaneous running queries.
This requires a Gafaelfawr token with `admin:userinfo` scope to ask
Gafaelfawr for the current quota for a user. Note that this means
the Qserv Kafka bridge will only work for Phalanx environments that
use LDAP as a user information source.