Update resources for Cloud SQL Proxy
The new version of Cloud SQL Proxy uses more memory, enough that
we're triggering memory alerts for some services. Increase the
requests and limits based on current graphs from idfint.

times-square was not correctly applying resource requests and
limits to its Cloud SQL Auth Proxy sidecar because of an indentation
error, also fixed in this change.