Switch to uv
Move all dependencies into `pyproject.toml` instead of using separate
dependency lists. Switch to `uv.lock` for dependency management
instead of frozen dependency files. Document uv as a new requirement
for developing Phalanx.

Move path analysis into a separate step of the GitHub Actions workflow
and simplify the conditionals to skip entire steps that do not apply
to a given pull request. Use that to merge the docs workflow into the
main CI workflow.

Move the versions of the various tools used in the GitHub Actions
workflow to variables set at the top of the workflow.