Also run minikube tests on uv.lock changes
We want to re-run the installer if the Python dependencies change.
This will have some false positives for development dependencies,
but we usually update with `make update` and it usually changes all
the dependencies, so it's probably worth the check.

Remove `environments/Chart.yaml` from the triggers since we should
basically never change that file.