Add Argo CD role:portal on idfdev and idfint
Add a new Argo CD role for Portal developers, although for the time
being don't add anyone to that role. If/when we grant lsst.cloud
accounts to the new Portal developers, they can be added to that
group.

Clean up the developer role a bit to remove stanzas that should be
unnecessary per the Argo CD documentation.