Move MDX content to a directory of files
Rather than embed the page MDX in the values.yaml, we're moving it into
the content/ directory of the Squareone chart. This will make
the content easier to manage as individual files rather than strings in
YAML.

Note that Files.Get reads paths relative to the chart root, so the
content directory is accessible from templates.