Read MDX content with Files.Get
This updates configmap-mdx.yaml so that it reads the MDX content from
the content directory. It attempts to get an environment-specific file
and then falls back to idfprod if none exists.

Note that Helm's YAML formatter seems to be doing some reformatting.
Namely the keys aren't being quoted, and a mix of > and | block syntax
is used in the rendered charts. The configmap still seems functional
though.

fixup use literal block syntax for mdx content

tweak template syntax