Upgrade Butler server to 4.1.2
Upgrade Butler server to a version with a fix for a forward-compatibility issue with a new version of the client (from DM-52397) that will be released in a couple weeks.  This also includes minor bugfixes for a few exceptions that should have been propagated to the client but were instead raising 500 server errors.