ingress-nginx - Attach correct demo GKE cluster ingress-nginx IP address
This contained the old demo cluster ingress address, though it seems to
have just ignored it and the new cluster ingress-nginx LoadBalancer
Service just made a new ephemeral Google external IP address for itself
and attached it just fine.

In the meantime, I (Dan) tried to convert the new working ephemeral IP
address into a static IP address so that I could get it covered by
Terraform config in `idf_deploy`... But I mistyped the name, and of
course this can't be changed. So I made a new Google static external IP
address with the correct name, and hopefully this change will attach it
to the demo cluster ingress. If it does, then I'll change DNS to match.