Create turborepo-cache application
The turborepo-remote-cache application from ductors is an open-source
implementation of Vercel's Turborepo remote caching service. This
Turborepo cache will speed up
[Squareone](https://github.com/lsst-sqre/squareone) CI and local
development by centralizing caching so that tests and builds are
unnecessarily repeated.

The turborepo-remote-cache works with a Google Cloud Storage bucket and
authenticates with workload identity. See
https://github.com/lsst/idf_deploy/pull/704 for set up.

For authentication, the turborepo-remote-cache can use either a JWT with
OIDC or a static token. For this deployment, I decided the cleanest
approach is to use a static token, but have a FastAPI proxy application
front the cache service so that we can effectively exchange a Gafaelfawr
user access token for the static token so that users don't need to know
about the static token. This proxy is the
https://github.com/lsst-sqre/turborepo-cache-proxy application.

At the moment we are co-opting the write:git-lfs scope for turborepo
usage but we'll create a custom scope just for turborepo in the future.