Update to qserv-kafka 4.0.0
The new version fixes a few bugs and updates the API version for
requests to Qserv if versioned API requests are enabled. Also drop
the application version from the labels for every Kubernetes object,
since that makes it harder to see what's really changed, and adjust
for the renaming of the `log_profile` configuration option.