Install sasquatch before other applications
Any apps that publishes app metrics need sasquatch to be synced
first. Otherwise, the apps will start in a state where they don't
publish metrics, and they will need to be restarted after sasquatch is
available.

There is a chicken-and-egg problem with gafaelfawr and sasquatch.
Gafaelfawr needs sasquatch to publish metrics, but sasquatch needs
gafaelfawr to provision GafaelfawrIngresses. Gafaelfawr will start
without sasquatch, but it must be restarted after sasquatch is available
so that it publishes metrics.